# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Dealna backend API",
        "title": "Dealna API",
        "contact": {
            "name": "Amro Deek"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/giveaway/queue/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all active queue entries for the authenticated user",
                "tags": [
                    "Giveaway Queue"
                ],
                "summary": "Get user's queue entries",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.QueuePosition"
                            }
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/giveaway/queue/{itemId}/entries": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all active queue entries for a specific item (Owner only)",
                "tags": [
                    "Giveaway Queue"
                ],
                "summary": "Get queue entries for an item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Item ID",
                        "name": "itemId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.QueueEntry"
                            }
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/giveaway/queue/{itemId}/entries/{entryId}/accept": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Owner accepts the current reserved user's turn",
                "tags": [
                    "Giveaway Queue"
                ],
                "summary": "Owner accepts turn",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Item ID",
                        "name": "itemId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Queue Entry ID",
                        "name": "entryId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/giveaway/queue/{itemId}/entries/{entryId}/complete": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Receiver confirms they received the item",
                "tags": [
                    "Giveaway Queue"
                ],
                "summary": "Receiver confirms handoff",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Item ID",
                        "name": "itemId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Queue Entry ID",
                        "name": "entryId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/giveaway/queue/{itemId}/entries/{entryId}/handoff": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Owner initiates the item handoff to the confirmed user",
                "tags": [
                    "Giveaway Queue"
                ],
                "summary": "Owner initiates handoff",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Item ID",
                        "name": "itemId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Queue Entry ID",
                        "name": "entryId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/giveaway/queue/{itemId}/entries/{entryId}/reject": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Owner rejects the current reserved user's turn and skips to next",
                "tags": [
                    "Giveaway Queue"
                ],
                "summary": "Owner rejects turn",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Item ID",
                        "name": "itemId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Queue Entry ID",
                        "name": "entryId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/giveaway/queue/{itemId}/join": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Adds the authenticated user to the giveaway queue for the specified item",
                "tags": [
                    "Giveaway Queue"
                ],
                "summary": "Join an item's giveaway queue",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Item ID",
                        "name": "itemId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.QueueEntry"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/giveaway/queue/{itemId}/leave": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes the authenticated user from the giveaway queue for the specified item",
                "tags": [
                    "Giveaway Queue"
                ],
                "summary": "Leave an item's giveaway queue",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Item ID",
                        "name": "itemId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/giveaway/queue/{itemId}/position/{entryId}": {
            "get": {
                "description": "Retrieves the current numerical position of a specific queue entry for an item",
                "tags": [
                    "Giveaway Queue"
                ],
                "summary": "Get user position in queue",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Item ID",
                        "name": "itemId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Queue Entry ID",
                        "name": "entryId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns { \\\"position\\\": 1 }",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.QueueEntry": {
            "type": "object",
            "properties": {
                "buyerName": {
                    "type": "string"
                },
                "buyerPic": {
                    "type": "string"
                },
                "entryID": {
                    "type": "string"
                },
                "entryStatus": {
                    "$ref": "#/definitions/domain.QueueEntryStatus"
                },
                "itemID": {
                    "type": "string"
                },
                "joinedAt": {
                    "type": "string"
                },
                "turnStartedAt": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
        "domain.QueueEntryStatus": {
            "type": "string",
            "enum": [
                "WAITING",
                "RESERVED",
                "CONFIRMED",
                "HANDED_OFF",
                "EXPIRED",
                "COMPLETED",
                "CANCELLED"
            ],
            "x-enum-varnames": [
                "QueueStatusWaiting",
                "QueueStatusReserved",
                "QueueStatusConfirmed",
                "QueueStatusHandedOff",
                "QueueStatusExpired",
                "QueueStatusCompleted",
                "QueueStatusCancelled"
            ]
        },
        "domain.QueuePosition": {
            "type": "object",
            "properties": {
                "entry": {
                    "$ref": "#/definitions/domain.QueueEntry"
                },
                "position": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}